+++ rq/utils.py
@@ -14,10 +14,11 @@ import importlib
 import logging
 import numbers
 import sys
-try:
-    from collections.abc import Iterable
-except ImportError:
-    from collections import Iterable
+
+from collections.abc import Iterable
+from distutils.version import StrictVersion
+
+from redis.exceptions import ResponseError
 
 from .compat import as_text, is_python_version, string_types
 from .exceptions import TimeoutFormatError
@@ -251,3 +252,15 @@ def parse_timeout(timeout):
                                          'such as "1h", "23m".')
 
     return timeout
+
+
+def get_version(connection):
+    """
+    Returns StrictVersion of Redis server version.
+    This function also correctly handles 4 digit redis server versions.
+    """
+    try:
+        version_string = connection.info("server")["redis_version"]
+    except ResponseError:  # fakeredis doesn't implement Redis' INFO command
+        version_string = "5.0.9"
+    return StrictVersion('.'.join(version_string.split('.')[:3]))